feat: add calendar export options for hot desk bookings
