+++ src/app/reserve/[venueId]/reservation-client.tsx
@@ -10,7 +10,11 @@ import {
   Sparkles,
   UsersRound,
   Zap,
+  CalendarPlus,
+  Mail,
+  Download,
 } from "lucide-react";
+import { getCalendarUrls, downloadICS } from "@/lib/calendar";
 
 type Seat = {
   id: string;
@@ -52,6 +56,7 @@ export default function ReservationClient({ venue }: { venue: Venue }) {
   const [loading, setLoading] = useState(true);
   const [booking, setBooking] = useState(false);
   const [message, setMessage] = useState("");
+  const [confirmationId, setConfirmationId] = useState<string | null>(null);
 
   const loadAvailability = useCallback(async () => {
     setLoading(true);
@@ -150,6 +155,7 @@ export default function ReservationClient({ venue }: { venue: Venue }) {
 
     setBooking(true);
     setMessage("");
+    setConfirmationId(null);
 
     const response = await fetch("/api/reservations/book", {
       method: "POST",
@@ -176,6 +182,7 @@ export default function ReservationClient({ venue }: { venue: Venue }) {
     setMessage(
       `${selected.seatNumber} confirmed. Reference: ${payload.confirmationId}`,
     );
+    setConfirmationId(payload.confirmationId);
     setSelectedSeat(null);
     setBooking(false);
     await loadAvailability();
@@ -200,8 +207,34 @@ export default function ReservationClient({ venue }: { venue: Venue }) {
         </header>
 
         {message && (
-          <div className="mb-6 rounded-2xl border border-violet-400/20 bg-violet-400/10 px-4 py-3 text-sm text-violet-100">
-            {message}
+          <div className="mb-6 rounded-2xl border border-violet-400/20 bg-violet-400/10 p-5 text-sm text-violet-100">
+            <p className="font-semibold">{message}</p>
+            {confirmationId && (
+              <div className="mt-4 flex flex-wrap items-center gap-3">
+                <a
+                  href={getCalendarUrls(venue.name, venue.address || "", date, time).googleUrl}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="flex items-center gap-2 rounded-xl bg-violet-600/20 border border-violet-500/30 px-4 py-2 hover:bg-violet-600/40 transition-colors"
+                >
+                  <CalendarPlus className="h-4 w-4" /> Add to Google
+                </a>
+                <a
+                  href={getCalendarUrls(venue.name, venue.address || "", date, time).outlookUrl}
+                  target="_blank"
+                  rel="noopener noreferrer"
+                  className="flex items-center gap-2 rounded-xl bg-violet-600/20 border border-violet-500/30 px-4 py-2 hover:bg-violet-600/40 transition-colors"
+                >
+                  <Mail className="h-4 w-4" /> Add to Outlook
+                </a>
+                <button
+                  onClick={() => downloadICS(venue.name, venue.address || "", date, time)}
+                  className="flex items-center gap-2 rounded-xl bg-white/5 border border-white/10 px-4 py-2 hover:bg-white/10 transition-colors text-zinc-300"
+                >
+                  <Download className="h-4 w-4" /> Download .ics
+                </button>
+              </div>
+            )}
           </div>
         )}
